fix stylesЭ
