+++ src/main/java/ok/dht/test/skroba/dao/MemorySegmentDao.java
@@ -1,31 +1,25 @@
 package ok.dht.test.skroba.dao;
 
-
 import jdk.incubator.foreign.MemorySegment;
 import ok.dht.test.skroba.dao.base.Config;
 import ok.dht.test.skroba.dao.base.Dao;
 import ok.dht.test.skroba.dao.base.Entry;
 import ok.dht.test.skroba.dao.comparators.EntryKeyComparator;
-import ok.dht.test.skroba.dao.comparators.MemorySegmentComparator;
 import ok.dht.test.skroba.dao.exceptions.TooManyFlushesInBgException;
 import ok.dht.test.skroba.dao.iterators.MergeIterator;
+import ok.dht.test.skroba.dao.iterators.TombstoneFilteringIterator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collection;
 import java.util.Iterator;
-import java.util.NoSuchElementException;
+import java.util.List;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentSkipListMap;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicLong;
 import java.util.concurrent.locks.ReadWriteLock;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 
@@ -39,14 +33,12 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
     
     private final ExecutorService executor = Executors.newSingleThreadExecutor(
             r -> new Thread(r, "MemorySegmentDaoBG"));
-    
-    private volatile State state;
-    
     private final Config config;
+    private volatile State stateOfClass;
     
     public MemorySegmentDao(Config config) throws IOException {
         this.config = config;
-        this.state = State.newState(config, Storage.load(config));
+        this.stateOfClass = State.newState(config, StorageUtils.load(config));
     }
     
     @Override
@@ -61,7 +53,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
     private TombstoneFilteringIterator getTombstoneFilteringIterator(MemorySegment from, MemorySegment to) {
         State state = accessState();
         
-        ArrayList<Iterator<Entry<MemorySegment>>> iterators = state.storage.iterate(from, to);
+        List<Iterator<Entry<MemorySegment>>> iterators = state.storage.iterate(from, to);
         
         iterators.add(state.flushing.get(from, to));
         iterators.add(state.memory.get(from, to));
@@ -114,7 +106,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
             }
             
             state = state.prepareForFlush();
-            this.state = state;
+            this.stateOfClass = state;
         } finally {
             upsertLock.writeLock().unlock();
         }
@@ -124,12 +116,12 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
                 State state = accessState();
                 
                 Storage storage = state.storage;
-                Storage.save(config, storage, state.flushing.values());
-                Storage load = Storage.load(config);
+                StorageUtils.save(config, storage, state.flushing.values());
+                Storage load = StorageUtils.load(config);
                 
                 upsertLock.writeLock().lock();
                 try {
-                    this.state = state.afterFlush(load);
+                    this.stateOfClass = state.afterFlush(load);
                 } finally {
                     upsertLock.writeLock().unlock();
                 }
@@ -138,7 +130,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
             } catch (Exception e) {
                 LOG.error("Can't flush", e);
                 try {
-                    this.state.storage.close();
+                    this.stateOfClass.storage.close();
                 } catch (IOException ex) {
                     LOG.error("Can't stop storage", ex);
                     ex.addSuppressed(e);
@@ -155,7 +147,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
         // it is intentionally the read lock!!!
         upsertLock.writeLock().lock();
         try {
-            runFlush = state.memory.overflow();
+            runFlush = stateOfClass.memory.overflow();
         } finally {
             upsertLock.writeLock().unlock();
         }
@@ -181,7 +173,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
                 return null;
             }
             
-            Storage.compact(
+            StorageUtils.compact(
                     config,
                     () -> MergeIterator.of(
                             state.storage.iterate(VERY_FIRST_KEY,
@@ -191,11 +183,11 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
                     )
             );
             
-            Storage storage = Storage.load(config);
+            Storage storage = StorageUtils.load(config);
             
             upsertLock.writeLock().lock();
             try {
-                this.state = state.afterCompact(storage);
+                this.stateOfClass = state.afterCompact(storage);
             } finally {
                 upsertLock.writeLock().unlock();
             }
@@ -224,7 +216,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
     }
     
     private State accessState() {
-        State state = this.state;
+        State state = this.stateOfClass;
         if (state.closed) {
             throw new IllegalStateException("Dao is already closed");
         }
@@ -233,7 +225,7 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
     
     @Override
     public synchronized void close() throws IOException {
-        State state = this.state;
+        State state = this.stateOfClass;
         if (state.closed) {
             return;
         }
@@ -244,194 +236,12 @@ public class MemorySegmentDao implements Dao<MemorySegment, Entry<MemorySegment>
         } catch (InterruptedException e) {
             throw new IllegalStateException(e);
         }
-        state = this.state;
+        state = this.stateOfClass;
         state.storage.close();
-        this.state = state.afterClosed();
+        this.stateOfClass = state.afterClosed();
         if (state.memory.isEmpty()) {
             return;
         }
-        Storage.save(config, state.storage, state.memory.values());
+        StorageUtils.save(config, state.storage, state.memory.values());
     }
-    
-    private static class TombstoneFilteringIterator implements Iterator<Entry<MemorySegment>> {
-        private final Iterator<Entry<MemorySegment>> iterator;
-        private Entry<MemorySegment> current;
-        
-        public TombstoneFilteringIterator(Iterator<Entry<MemorySegment>> iterator) {
-            this.iterator = iterator;
-        }
-        
-        public Entry<MemorySegment> peek() {
-            return hasNext() ? current : null;
-        }
-        
-        @Override
-        public boolean hasNext() {
-            if (current != null) {
-                return true;
-            }
-            
-            while (iterator.hasNext()) {
-                Entry<MemorySegment> entry = iterator.next();
-                if (!entry.isTombstone()) {
-                    this.current = entry;
-                    return true;
-                }
-            }
-            
-            return false;
-        }
-        
-        @Override
-        public Entry<MemorySegment> next() {
-            if (!hasNext()) {
-                throw new NoSuchElementException("...");
-            }
-            Entry<MemorySegment> next = current;
-            current = null;
-            return next;
-        }
-    }
-    
-    private static class State {
-        final Config config;
-        final Memory memory;
-        final Memory flushing;
-        final Storage storage;
-        final boolean closed;
-        
-        State(Config config, Memory memory, Memory flushing, Storage storage) {
-            this.config = config;
-            this.memory = memory;
-            this.flushing = flushing;
-            this.storage = storage;
-            this.closed = false;
-        }
-        
-        State(Config config, Storage storage, boolean closed) {
-            this.config = config;
-            this.memory = Memory.EMPTY;
-            this.flushing = Memory.EMPTY;
-            this.storage = storage;
-            this.closed = closed;
-        }
-        
-        static State newState(Config config, Storage storage) {
-            return new State(
-                    config,
-                    new Memory(config.flushThresholdBytes()),
-                    Memory.EMPTY,
-                    storage
-            );
-        }
-        
-        public State prepareForFlush() {
-            checkNotClosed();
-            if (isFlushing()) {
-                throw new IllegalStateException("Already flushing");
-            }
-            return new State(
-                    config,
-                    new Memory(config.flushThresholdBytes()),
-                    memory,
-                    storage
-            );
-        }
-        
-        public State afterFlush(Storage storage) {
-            checkNotClosed();
-            if (!isFlushing()) {
-                throw new IllegalStateException("Wasn't flushing");
-            }
-            return new State(
-                    config,
-                    memory,
-                    Memory.EMPTY,
-                    storage
-            );
-        }
-        
-        public State afterCompact(Storage storage) {
-            checkNotClosed();
-            return new State(
-                    config,
-                    memory,
-                    flushing,
-                    storage
-            );
-        }
-        
-        public State afterClosed() {
-            checkNotClosed();
-            if (!storage.isClosed()) {
-                throw new IllegalStateException("Storage should be closed early");
-            }
-            return new State(config, storage, true);
-        }
-        
-        public void checkNotClosed() {
-            if (closed) {
-                throw new IllegalStateException("Already closed");
-            }
-        }
-        
-        public boolean isFlushing() {
-            return this.flushing != Memory.EMPTY;
-        }
-    }
-    
-    private static class Memory {
-        
-        static final Memory EMPTY = new Memory(-1);
-        private final AtomicLong size = new AtomicLong();
-        private final AtomicBoolean oversized = new AtomicBoolean();
-        
-        private final ConcurrentSkipListMap<MemorySegment, Entry<MemorySegment>> delegate =
-                new ConcurrentSkipListMap<>(MemorySegmentComparator.INSTANCE);
-        
-        private final long sizeThreshold;
-        
-        Memory(long sizeThreshold) {
-            this.sizeThreshold = sizeThreshold;
-        }
-        
-        public boolean isEmpty() {
-            return delegate.isEmpty();
-        }
-        
-        public Collection<Entry<MemorySegment>> values() {
-            return delegate.values();
-        }
-        
-        public boolean put(MemorySegment key, Entry<MemorySegment> entry) {
-            if (sizeThreshold == -1) {
-                throw new UnsupportedOperationException("Read-only map");
-            }
-            Entry<MemorySegment> segmentEntry = delegate.put(key, entry);
-            long sizeDelta = Storage.getSizeOnDisk(entry);
-            if (segmentEntry != null) {
-                sizeDelta -= Storage.getSizeOnDisk(segmentEntry);
-            }
-            long newSize = size.addAndGet(sizeDelta);
-            if (newSize > sizeThreshold) {
-                return !oversized.getAndSet(true);
-            }
-            return false;
-        }
-        
-        public boolean overflow() {
-            return !oversized.getAndSet(true);
-        }
-        
-        public Iterator<Entry<MemorySegment>> get(MemorySegment from, MemorySegment to) {
-            return to == null
-                    ? delegate.tailMap(from).values().iterator()
-                    : delegate.subMap(from, to).values().iterator();
-        }
-        
-        public Entry<MemorySegment> get(MemorySegment key) {
-            return delegate.get(key);
-        }
-    }
-    
 }
